refactor: move getEnvironmentVariable to separate module
мелочь
